Split up the Helper class into ErrorHandler and WithdrawHelper
