+++ src/main/kotlin/net/trueog/diamondbankog/WithdrawHelper.kt
@@ -1,13 +1,12 @@
 package net.trueog.diamondbankog
 
+import net.trueog.diamondbankog.ErrorHandler.handleError
 import net.trueog.diamondbankog.InventoryExtensions.withdraw
-import net.trueog.diamondbankog.PostgreSQL.PlayerShards
 import net.trueog.diamondbankog.PostgreSQL.ShardType
 import org.bukkit.entity.Player
-import java.util.*
 import kotlin.math.floor
 
-object Helper {
+object WithdrawHelper {
     /**
      * @return The amount of not removed shards, -1 if error
      */
@@ -116,42 +115,5 @@ object Helper {
         )
 
         return notRemovedInventory + notRemovedEnderChest
-    }
-
-    class EconomyException(message: String) : Exception(message)
-
-    /**
-     * Handles the error by throwing, disables the economy unless you specify it shouldn't
-     */
-    fun handleError(
-        uuid: UUID,
-        expectedMutatedShards: Int,
-        playerShards: PlayerShards?,
-        otherUuid: UUID? = null,
-        dontDisableEconomy: Boolean = false
-    ) {
-        if (!dontDisableEconomy) DiamondBankOG.economyDisabled = true
-
-        throw EconomyException(
-            """
-
-            Player UUID: $uuid
-            ${
-                if (otherUuid != null) "Other Player UUID: $otherUuid" else ""
-            }Expected Mutated Shards = $expectedMutatedShards${
-                if (playerShards != null) {
-                    if (playerShards.shardsInBank != -1) "Player Bank Balance: ${playerShards.shardsInBank}" else ""
-                } else ""
-            }${
-                if (playerShards != null) {
-                    if (playerShards.shardsInInventory != -1) "Player Inventory Balance: ${playerShards.shardsInInventory}" else ""
-                } else ""
-            }${
-                if (playerShards != null) {
-                    if (playerShards.shardsInEnderChest != -1) "Player Ender Chest Balance: ${playerShards.shardsInEnderChest}" else ""
-                } else ""
-            }
-        """.trimIndent()
-        )
     }
 }
